refactor: update the paths
