+++ src/lib/substrate/index.ts
@@ -1,4 +1,8 @@
-import { SubstrateProvider, useSubstrate, useSubstrateState } from './substrate-context';
+import { SubstrateProvider, useSubstrate, useSubstrateSecureState, useSubstrateState } from './substrate-context';
 import utils from './utils';
 
-export { SubstrateProvider, utils as substrateUtils, useSubstrate, useSubstrateState };
+// ray test touch <
+export * from './context/types';
+// ray test touch >
+
+export { SubstrateProvider, utils as substrateUtils, useSubstrate, useSubstrateSecureState, useSubstrateState };
